workers/ai-infrastructure/requirements.txt
praw==7.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
